requirements.txt
# Core Python dependencies
python-dotenv>=0.19.0
requests>=2.28.0
aiohttp>=3.8.0
pandas>=1.5.0
numpy>=1.21.0
oauthlib>=3.2.0

# Yahoo Fantasy Wrapper dependencies
yahoo-fantasy-api>=1.0.0
oauthlib>=3.2.0
requests-oauthlib>=1.3.1

# Fantasy AI Ultimate dependencies
fastapi>=0.95.0
uvicorn>=0.20.0
sqlalchemy>=1.4.0
alembic>=1.8.0
psycopg2-binary>=2.9.0
redis>=4.3.0

# AI and ML dependencies (optional for initial setup)
# scikit-learn>=1.1.0
# tensorflow>=2.10.0
# torch>=1.12.0
# transformers>=4.20.0

# Data processing
pydantic>=1.9.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
aiofiles>=0.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.23.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.991 